F6. Nations in North America converted from native_council are in the interregnum state, as they do not have ruler specified.
For every despotic_monarchy that was having "Native Council" as ruler,
added a new rule with ruler name randomly selected from male_names of culture.
Dynasty and age is left to be choosen by game. The inuguration date is allways
1444.11.11.
Made the ruler stat more random, instead of always 3,3,3.
This is determined as follows:
 - choose sum of stats (5% for 6,10% for 7, 70% for 8, 10% for 9, 5% for 10)
 - distribute it randomly across categories