F1. Nations in High American technology group sees (not covered by TI)     only provinces that are owned by this nations.
Generated as list of the provinces that are discovered in vanilla
by one of following technology groups: south_american, andean,
mesoamerican, north_american and performed set difference with
list of provinces that are discovered by high_american in SI conversion.
For each province that was left, province_no.txt files was generated
with "discovered_by = high_american" content.
Some of this provinces overlap with list of provinces that was
discovered by Aztec only for SI cost, so same files get overridden,
however following duplicate entries shall be removed in case
of merge:
 * 1508 - SI Coast.txt
 * 1509 - SI Coast.txt
 * 1510 - SI Coast.txt
 * 1550 - SI Coast.txt
 * 1552 - SI Coast.txt
 * 1553 - SI Coast.txt